+++ api/src/User/Entity/User.php
@@ -1232,6 +1232,12 @@ class User implements UserInterface, EquatableInterface
      */
     private $unreadSolidaryMessageNumber;
 
+    /**
+     * @var int|null The savedCo2 of this user in grams
+     * @Groups({"readUser","results","write", "threads", "thread", "readCommunity", "readCommunityUser", "readEvent", "massMigrate", "readExport","readPublicProfile","readReview"})
+     */
+    private $savedCo2;
+
     public function __construct($status = null)
     {
         $this->id = self::DEFAULT_ID;
@@ -2928,6 +2934,18 @@ class User implements UserInterface, EquatableInterface
     public function setUnreadSolidaryMessageNumber(?int $unreadSolidaryMessageNumber): self
     {
         $this->unreadSolidaryMessageNumber = $unreadSolidaryMessageNumber;
+
+        return $this;
+    }
+
+    public function getSavedCo2(): ?int
+    {
+        return $this->savedCo2;
+    }
+
+    public function setSavedCo2(?int $savedCo2): self
+    {
+        $this->savedCo2 = $savedCo2;
 
         return $this;
     }